+++ js/views/background.js
@@ -1,27 +1,28 @@
 'use strict';
 
 import { setBadgeText } from '../app/browser.js';
-import { ListingPoller } from '../app/manager/listingpoller.js';
-import { browserLocalStorage, onMessage } from '../app/browser.js';
+import { ListingWorker } from '../app/manager/listingworker.js';
+import { onMessage } from '../app/browser.js';
 
-const listingPoller = new ListingPoller();
+const LOAD_LISTINGS_ALARM_KEY = 'load-listings';
+const listingWorker = new ListingWorker();
 
 function addListeners() {
-    onMessage.addListener((request, _sender, sendResponse) => {
-        switch (request.name) {
+    onMessage.addListener(({ name }, _sender, sendResponse) => {
+        switch (name) {
             case 'startLoading': {
                 // force load
-                listingPoller.resumeLoading(true);
+                load(true);
                 sendResponse();
                 break;
             }
             case 'resumeLoading': {
-                listingPoller.resumeLoading();
+                load();
                 sendResponse();
                 break;
             }
             case 'clearListingCount': {
-                listingPoller.clearListingCount();
+                listingWorker.clearListingCount();
                 updateCount(0);
                 sendResponse();
                 break;
@@ -43,15 +44,25 @@ function addListeners() {
         });
     });
     
-    listingPoller.on('count', (count) => {
+    listingWorker.on('count', (count) => {
         updateCount(count);
+    });
+    
+    chrome.alarms.onAlarm.addListener((alarm) => {
+        switch (alarm.name) {
+            case LOAD_LISTINGS_ALARM_KEY: {
+                load();
+                break;
+            }
+            default: {
+                console.warn('Unknown alarm:', alarm.name);
+            }
+        }
     });
 }
 
 // updates the count on using the badge text
 function updateCount(count) {
-    browserLocalStorage.setItem('listingCount', count);
-    
     if (count >= 1000) {
         // truncate to fit
         count = '999+';
@@ -67,9 +78,44 @@ function updateCount(count) {
     });
 }
 
+async function checkAlarmState() {
+    const alarm = await chrome.alarms.get(LOAD_LISTINGS_ALARM_KEY);
+    
+    if (!alarm) {
+        // 1 minute - don't call it immediately
+        startAlarm(1);
+    }
+}
+
+async function startAlarm(delayInMinutes) {
+    await chrome.alarms.create(LOAD_LISTINGS_ALARM_KEY, {
+        delayInMinutes
+    });
+}
+
+async function load(force = false) {
+    if (listingWorker.isLoading) {
+        // do nothing
+        return;
+    }
+    
+    async function next() {
+        const pollIntervalMinutes = await listingWorker.getPollIntervalMinutes();
+        
+        await startAlarm(pollIntervalMinutes);
+    }
+    
+    listingWorker.start(force)
+        .then(next)
+        .catch((err) => {
+            console.warn('Error getting listings:', err);
+            next();
+        });
+}
+
 // ready
 {
     addListeners();
     updateCount(0);
-    listingPoller.start(5);
+    checkAlarmState();
 }